requirements.txt
av
py-nats
aiohttp